requirements.server.txt
pytorch-lightning
learn2learn >= 0.1.6
comet-ml
resemblyzer
g2p-en
inflect
librosa
matplotlib
numba
numpy
pypinyin
PyYAML
scikit-learn
scipy
soundfile
tensorboard
tgt
tqdm
unidecode
pandas
pyloudnorm
google-cloud-storage==1.43.0
fastapi==0.85.1
uvicorn==0.16.0
smart_open==6.2.0
